requirements.txt
wheel
Babel
gdal
click>7,<=9
Flask
jsonschema
pyproj
python-dateutil
pytz
PyYAML
py_linq
rasterio
shapely
tinydb
unicodecsv
future-annotations
pydantic
jinja2==3.0.3
xarray
